When checking in more than once the user is redirected to an error page

+++ src/checkin.php
@@ -3,15 +3,19 @@
 include_once('Database/DatabaseConnection.php');
 
 if (isset($_GET['id'])) {
-
     $ticketId = $_GET['id'];
     $pdo = DatabaseConnection::getInstance();
-    // TODO: try-catch
-    $stmt = $pdo->prepare("update tickets set checkin = 1 where id=:id");
+    $stmt = $pdo->prepare("select * from tickets where id = :id");
     $stmt->execute(['id' => $ticketId]);
+    $checkedIn = $stmt->fetch()['checkin'];
 
-    unlink("output/$ticketId.pdf");
-
-    header('Location: /thank_you.php');
-
+    if ($checkedIn != 1) {
+        $stmt = $pdo->prepare("update tickets set checkin = 1 where id=:id");
+        $stmt->execute(['id' => $ticketId]);
+        unlink("output/$ticketId.pdf");
+        header('Location: /thank_you.php');
+    } else {
+        http_response_code(400);
+        header("Location: checkin_error.php");
+    }
 }
